requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
async-timeout==5.0.1
attrs==25.3.0
catboost==1.2.8
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
einops==0.8.1
et-xmlfile==2.0.0
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
graphviz==0.20.3
greenlet==3.1.1
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
mako==1.3.10
markupsafe==3.0.2
mpmath==1.3.0
multidict==6.3.2
narwhals==1.35.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8 
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openpyxl==3.1.5
optuna==3.5.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
propcache==0.3.1
psutil==7.0.0
pycparser==2.22
pyparsing==3.2.1
pyqt5==5.15.11
pyqt5-qt5==5.15.16
pyqt5-sip==12.17.0
pyside2==5.15.2.1
pyside6==6.9.0
pyside6-addons==6.9.0
pyside6-essentials==6.9.0
python-dateutil==2.9.0.post0
pytz==2025.1
pywavelets==1.8.0
pyyaml==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
shap==0.47.1
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
soundfile==0.13.1
sqlalchemy==2.0.40
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0
torch-geometric==2.6.1
tqdm==4.67.1
triton==3.2.0
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wfdb==4.2.0
wheel==0.45.1
yarl==1.18.3